requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via starlette
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
fastapi==0.115.6
    # via -r requirements.in
google-api-core==2.24.0
    # via google-api-python-client
google-api-python-client==2.156.0
    # via -r requirements.in
google-auth==2.37.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.66.0
    # via google-api-core
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via
    #   anyio
    #   requests
oauthlib==3.2.2
    # via requests-oauthlib
proto-plus==1.25.0
    # via google-api-core
protobuf==5.29.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.4
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
pyparsing==3.2.0
    # via httplib2
python-dotenv==1.0.1
    # via pydantic-settings
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.in
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
